% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_gene_expr.R
\name{boxplot_gene_expr}
\alias{boxplot_gene_expr}
\title{boxplot_gene_expr}
\usage{
boxplot_gene_expr(data, gene)
}
\arguments{
\item{data}{data frame containing genes expression.}

\item{gene}{name of gene which expression we want to visualise.}
}
\value{
boxplot of expression.
}
\description{
Function \code{boxplot_gene_expr} generate a boxplot of expression for choosen gene.
}
\seealso{
genereg_vs_met
}
