% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_that.R
\name{boxplot_that}
\alias{boxplot_that}
\title{boxplot_that}
\usage{
boxplot_that(data, column, condition = "")
}
\arguments{
\item{data}{data frame containing interesing values.}

\item{column}{string containing name of column with values for boxplot.}

\item{condition}{vector of length equal to numer of rows of data,
\code{condition} should contains names of groups corresponding to rows.}
}
\value{
Object of class ggplot containing boxplot for values from choosen column.
}
\description{
Function \code{boxplot_that} generates a boxplot of values from choosen data frame column.
}
