% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_methylation_all_surv.R
\docType{data}
\name{BRCA_methylation_all_surv}
\alias{BRCA_methylation_all_surv}
\title{Methylation levels for different patients with BRCA cancer.}
\format{data.frame}
\description{
Rows of this data set correspond to patients, more precisely, to samples taken from patients.
First columns correspond to a samples and subtypes of BRCA cancer, next columns correspond to CpG islands from all chromosomes.
Values inside the table indicate the methylation level of CpG island for specified sample.
}
