% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_data.R
\name{chromosome_data}
\alias{chromosome_data}
\title{chromosome_data}
\usage{
chromosome_data(data, number)
}
\arguments{
\item{data}{methylation dataset}

\item{number}{number of chromosome}
}
\value{
dataset containing CpG islands from all methylation data.
}
\description{
Function \code{chromosome_data} produces a dataset for chosen chromosome from methylation data.
}
\examples{
chromosome_17 <- chromosome_data(BRCA_methylation_data_all, 17)

}
