% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_to_des.R
\name{con_to_des}
\alias{con_to_des}
\title{Function transforms 'condition' data format to 'design' data format}
\usage{
con_to_des(condition)
}
\arguments{
\item{condition}{vector of level values}
}
\value{
data frame of zeros and ones with rows coresponding to sample and columns coresponding to level
}
\description{
Transforming vector of levels corresponding to samples
}
