% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpG_mean.R
\name{CpG_mean}
\alias{CpG_mean}
\title{CpG_mean}
\usage{
CpG_mean(data, gene)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG islands, rows to samples.}

\item{gene}{name of choosen gene.}
}
\value{
Data frame containing
}
\description{
Function \code{CpG_mean} computes methylation means of CpG islands for choosen gene. In this case: “BRCA1”
}
