% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stats.R
\name{data_stats}
\alias{data_stats}
\title{Statistics for data}
\usage{
data_stats(data, condition)
}
\arguments{
\item{data}{dataset containing interesting values}

\item{condition}{vector of conditions}
}
\value{
data frame
}
\description{

}
