% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_plot.R
\name{em_plot}
\alias{em_plot}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
em_plot(data1, condition, names = NA)
}
\arguments{
\item{data1}{data.frame}

\item{condition}{condition fo groups}

\item{names}{condition fo groups}
}
\value{
plot
}
\description{
Function \code{p_values_plot} draws a plot with p-values from methylation and expression when we use the t-test. In this case we testing the hypothesis that the means in both groups of "cancer type" are equal.
}
\seealso{
vp_values_plot
}
