% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_data.R
\name{full_data}
\alias{full_data}
\title{Function creates a full set for metylation and expression which contains values from functions `mety_ttest` i `expr_nbinom`}
\usage{
full_data(dt1, dt2)
}
\arguments{
\item{dt1}{data.frame, a result of `expr_nbinom` function}

\item{dt2}{data.frame, a result of `mety_ttest` function}
}
\value{
data.frame which contains all statistics from both functions
}
\description{
Function creates a full set for metylation and expression which contains values from functions `mety_ttest` i `expr_nbinom`
}
