% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genereg_vs_met.R
\name{genereg_vs_met}
\alias{genereg_vs_met}
\title{genereg_vs_met}
\usage{
genereg_vs_met(data, gen)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG islands, rows to samples.}

\item{gen}{vector of levels coresponding to order of samples in data.}
}
\value{
A plot.
}
\description{
Function \code{genereg_vs_met} ...
}
