% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illumina_humanmethylation_27_data.R
\docType{data}
\name{illumina_humanmethylation_27_data}
\alias{illumina_humanmethylation_27_data}
\title{Illumina_humanmethylation_27_data}
\format{data.frame}
\description{
Illumina_humanmethylation_27_data
}
