% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islands_locations.R
\name{islands_locations}
\alias{islands_locations}
\title{Locations of CpG islands}
\usage{
islands_locations(data_A, data_B)
}
\arguments{
\item{data_A}{d}

\item{data_B}{d}
}
\value{
data frame
}
\description{
ddd
}
