% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_gene.R
\name{map_to_gene}
\alias{map_to_gene}
\title{Mapping CpG probes to genes}
\usage{
map_to_gene(data)
}
\arguments{
\item{data}{data frame  containing methylation values for CpG probes. Columns corresponds to probes, rows to samples.}
}
\value{
A data frame with CpG probes mapped to genes. If there were more than one probe corresponding to a gene, value is a mean of those probes.
}
\description{
Function \code{map_to_gene} maps CpG probes to corresponding genes.
}
\examples{
\dontrun{
BRCA_genes <- map_to_gene(BRCA_methylation_all)
}
}
