% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mety_ttest.R
\name{mety_ttest}
\alias{mety_ttest}
\title{mety_ttest: t test for metylation per each gene.}
\usage{
mety_ttest(data, condition)
}
\arguments{
\item{data}{data frame of zeros and ones. Rows coresponds to genes, columns to samples.}

\item{condition}{vector of levels coresponding to order of samples in data.}
}
\value{
A data frame with the following columns:
 \item{met.log.fold}{The log of the fold change.}
 \item{met.mean}{The base mean.}
 \item{met.t.stat}{The value of the t-test statistic.}
 \item{met.pval}{The p value for rejecting the null hypothesis.}
 \item{met.padj}{The adjusted p values}
 \item{met.B}{something...}
 \item{id}{he ID of the observable, taken from the row names of the counts slots.}
}
\description{
Function \code{mety_ttest} computes statistics and p-values for metylation data mapped to genes.
}
