% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mety_ttest2.R
\name{mety_ttest2}
\alias{mety_ttest2}
\title{t test for methylation per each gene from methyAnalysis package}
\usage{
mety_ttest2(data, condition)
}
\arguments{
\item{data}{object of class MehtyGenoSet, consisting methylation for each CpG island.}

\item{condition}{vector of levels coresponding to order of samples in data.}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{name}{The CpG probe name.}
 \item{mean}{The base mean.}
 \item{log2.fold}{The 2-log of the fold change.}
 \item{pval}{The p-values for rejecting the null hypothesis about the means equality.}
 \item{padj}{The adjusted p-values.}
}
\description{
Function \code{mety_ttest2} computes statistics and p-values for methylation data.
}
\seealso{
mety_ttest
}
