% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values_plot.R
\name{p_values_plot}
\alias{p_values_plot}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
p_values_plot(dt_expr, dt_met, names = NA, exp.pval, met.pval, id)
}
\arguments{
\item{dt_expr}{data.frame, a result of `expr_nbinom` function.}

\item{dt_met}{data.frame, a result of `mety_ttest` function.}

\item{names}{p-value below which...}

\item{exp.pval}{p-value for expression.}

\item{met.pval}{p-value for methylation.}

\item{id}{vector of genes symbols.}
}
\value{
plot
}
\description{
Function \code{p_values_plot} draws a plot with p-values from methylation and expression when we use the t-test. In this case we testing the hypothesis that the means in both groups of "cancer type" are equal.
}
\seealso{
volcano_plot
}
