% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generate.R
\name{report_generate}
\alias{report_generate}
\title{G}
\usage{
report_generate(data.e, data.m, condition.e, condition.m, test.e, test.m, genes)
}
\arguments{
\item{data.e}{data for expression}

\item{data.m}{data for methylation}

\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{test.e}{test results for expression}

\item{test.m}{test results for methylation}

\item{gene}{gene name}
}
\value{
pdf
}
\description{
Function \code{G}
}
