% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{data.fra}
}
\value{
plot
}
\description{
Function \code{volcano_plot} draws a plot with p-values and fold logarithm from methylation or expression when we use the t-test.
}
