% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_gen.R
\name{sum_gen}
\alias{sum_gen}
\title{boxplot_gene_expr}
\usage{
sum_gen(data, condition, gen)
}
\arguments{
\item{data}{data frame containing genes expression.}

\item{condition}{name of gene which expression we want to visualise.}

\item{gen}{s}
}
\value{
boxplot of expression.
}
\description{
Function \code{boxplot_gene_expr} generate a boxplot of expression for choosen gene.
}
\seealso{
genereg_vs_met
}
