% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diff.R
\name{test_diff}
\alias{test_diff}
\title{test_diff}
\usage{
test_diff(data, condition, test = "ttest")
}
\arguments{
\item{data}{data frame of ... Columns corresponds to genes, rows to samples.}

\item{condition}{vector of levels coresponding to order of samples in data.}

\item{test}{values: ttest, nbinom.}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{exp.mean}{The base mean.}
 \item{exp.mean.gr1}{The base mean for the counts for the first condition.}
 \item{exp.mean.g2}{The base mean for the second condition.}
 \item{exp.fold}{The ratio mean1/mean2.}
 \item{exp.log.fold}{The log2 of the fold change.}
 \item{exp.pval}{The p value for rejecting the null hypothesis 'mean1==mean2'}
 \item{exp.padj}{The adjusted p values (adjusted with 'p.adjust( pval, method="BH")')}
}
\description{
Function \code{tet_diff} computes statistics and p-values for choosen test.
}
