% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_met_brca.R
\docType{data}
\name{test_met_brca}
\alias{test_met_brca}
\title{Dataset from t-test.}
\format{data.frame}
\description{
\itemize{
  \item id gene symbol
  \item mean mean exression on gene
  \item log2.fold logarithm of fold 
  \item pval f
  \item padj f
}
}
