% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_gene.R
\name{visual_gene}
\alias{visual_gene}
\title{visual}
\usage{
visual_gene(condition.e, condition.m, data.m, data.e, gene, test.e, test.m)
}
\arguments{
\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{data.m}{data for methylation}

\item{data.e}{data for expression}

\item{gene}{gene name}

\item{test.e}{test results for expression}

\item{test.m}{test results for methylation}
}
\value{
A plot of class ggplot.
}
\description{
Function ...
}
