% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_volc.R
\name{visual_volc}
\alias{visual_volc}
\title{visual}
\usage{
visual_volc(condition.e, condition.m, data.m, data.e, gene, test.e, test.m)
}
\arguments{
\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{data.m}{data for methylation}

\item{data.e}{data for expression}

\item{gene}{gene name}

\item{list.test.e}{list of tests results for expression}

\item{list.test.m}{list of tests results for methylation}
}
\value{
A plot of class ggplot.
}
\description{
Function ...
}
