% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_volcano.R
\name{visual_volcano}
\alias{visual_volcano}
\title{Visulisations for methylation and expression.}
\usage{
visual_volcano(condition.e, condition.m, data.e, data.m, gene,
  test.e = list(), test.m = list())
}
\arguments{
\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{data.e}{data for expression}

\item{data.m}{data for methylation}

\item{gene}{gene name}

\item{test.e}{list of tests results for expression}

\item{test.m}{list of tests results for methylation}
}
\value{
A plot of class ggplot.
}
\description{
Function \code{visual_volcano} generate a dashboard with volcano plots for expression and methylation. Also it adds a tables with basic statistics.
}
\seealso{
volcano_plot, gene_stats
}
