% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
volcano_plot(data, line = NA, names = NA, ylog = TRUE, ngen = NA,
  title = NA, fold_line = NA, values = FALSE)
}
\arguments{
\item{data}{data.frame consisting result of chosen test}

\item{line}{p-value on which we draw a line.}

\item{names}{p-value below which...}

\item{ylog}{s}

\item{ngen}{symol or vector of gene names}

\item{title}{s}

\item{fold_line}{s}

\item{values}{logical value, TRUE if we want p-values and log fold for chosen gene}
}
\value{
plot
}
\description{
Function \code{volcano_plot} draws a plot with p-values and fold logarithm from methylation or expression when we use the t-test.
}
