% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpG_mean.R
\name{CpG_mean}
\alias{CpG_mean}
\title{CpG_mean}
\usage{
CpG_mean(data, gene)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG islands, rows to samples.}

\item{gene}{vector of levels coresponding to order of samples in data.}
}
\value{
A plot.
}
\description{
Function \code{CpG_mean} ...
}
