% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genereg_vs_met.R
\name{genereg_vs_met}
\alias{genereg_vs_met}
\title{genereg_vs_met (tylko w układzie współrzędnych HG18 a nie HG19)}
\usage{
genereg_vs_met(data, condition, gene, show_gen = FALSE, observ = FALSE)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG islands, rows to samples.}

\item{condition}{vextor of subtype of cancer}

\item{gene}{vector of levels coresponding to order of samples in data.}

\item{show_gen}{s}

\item{observ}{s}
}
\value{
A plot of class ggplot.
}
\description{
Function \code{genereg_vs_met} ...
}
