% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genereg_vs_met2.R
\name{genereg_vs_met2}
\alias{genereg_vs_met2}
\title{genereg_vs_met2}
\usage{
genereg_vs_met2(data, condition, gene, show_gen = FALSE)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG islands, rows to samples.}

\item{condition}{vextor of subtype of cancer}

\item{gene}{vector of levels coresponding to order of samples in data.}
}
\value{
A plot of class ggplot.
}
\description{
Function \code{genereg_vs_met2} ...
}
