% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_gene.R
\name{map_to_gene}
\alias{map_to_gene}
\title{mapping CpG islands to genes}
\usage{
map_to_gene(data)
}
\arguments{
\item{data}{data frame ... Columns corresponds to genes, rows to samples.}
}
\value{
A data frame with CpG islands mapped to genes. If there were more than one island corresponding to a gene, 
value is a mean of those islands.
}
\description{
Function \code{map_to_gene} maps CpG islands to corresponding genes.
}
