% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probes_locations.R
\name{probes_locations}
\alias{probes_locations}
\title{Locations of CpG probes}
\usage{
probes_locations(data, gene, condition)
}
\arguments{
\item{data}{d}

\item{gene}{d}

\item{condition}{d}
}
\value{
data frame
}
\description{
dfd
}
