% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_gene.R
\name{visual_gene}
\alias{visual_gene}
\title{Visualisations for genes.}
\usage{
visual_gene(condition.e, condition.m, data.e, data.m, gene, test.e, test.m)
}
\arguments{
\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{data.e}{data for expression}

\item{data.m}{data for methylation}

\item{gene}{gene name}

\item{test.e}{test results for expression}

\item{test.m}{test results for methylation}
}
\value{
A plot of class ggplot.
}
\description{
Function \code{visual_gene} generates a dashboard with methylation_path for methylation and boxplots for groups.
}
