% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
volcano_plot(dt_expr, dt_met, type, exp.log.fold, exp.pval, met.log.fold,
  met.pval, id)
}
\arguments{
\item{dt_expr}{data.frame, a result of `expr_nbinom` function.}

\item{dt_met}{data.frame, a result of `mety_ttest` function.}

\item{type}{we can choose a plot for methylation of expression.}

\item{exp.log.fold}{logarithm of expression fold.}

\item{exp.pval}{p-value for expression.}

\item{met.log.fold}{logarithm of methylation fold.}

\item{met.pval}{p-value for methylation.}

\item{id}{vector of genes symbols.}
}
\value{
plot
}
\description{
Function \code{volcano_plot} draws a plot with p-values and fold logarithm from methylation or expression when we use the t-test.
}
