% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
volcano_plot(data, line = NA, names = NA, log.fold, pval, id, ngen = NA,
  title = NA, fold_line = NA)
}
\arguments{
\item{data}{data.frame, a result of `expr_nbinom` function.}

\item{line}{p-value on which we draw a line.}

\item{names}{p-value below which...}

\item{log.fold}{logarithm of fold.}

\item{pval}{p-value.}

\item{id}{vector of genes symbols.}

\item{ngen}{s}

\item{title}{s}

\item{fold_line}{s}
}
\value{
plot
}
\description{
Function \code{volcano_plot} draws a plot with p-values and fold logarithm from methylation or expression when we use the t-test.
}
